pyproject.toml
[project]
name = "calendar-dingtalk-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.128.2",
    "httpx>=0.28.1",
    "icalendar>=6.3.2",
    "lxml>=6.0.2",
    "mcp[cli]>=1.2.0",
    "pydantic>=2.12.5",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.1",
    "uvicorn[standard]>=0.24.0",
]

[project.scripts]
calendar-dingtalk-client = "calendar_dingtalk_client:main"

[build-system]
requires = ["uv_build>=0.8.22,<0.9.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "black>=26.1.0",
    "mypy>=1.19.1",
    "pre-commit>=4.5.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "pytest-httpserver>=1.1.3",
    "ruff>=0.15.0",
    "types-python-dateutil>=2.9.0.20260124",
]
